+++ solidity/orders/src/Foo.sol
@@ -1,8 +1,0 @@
-// SPDX-License-Identifier: UNLICENSED
-pragma solidity >=0.8.25;
-
-contract Foo {
-    function id(uint256 value) external pure returns (uint256) {
-        return value;
-    }
-}