+++ spaceward/src/App.tsx
@@ -56,7 +56,7 @@ import { hashQueryKey } from "./utils/queryKeyHash.ts";
 import { DashboardPage } from "./pages/Dashboard.tsx";
 import { Web3OnboardProvider, init, useWagmiConfig } from "@web3-onboard/react";
 import injectedModule from "@web3-onboard/injected-wallets";
-import wagmi, { createConfig } from "@web3-onboard/wagmi";
+import wagmi, { createConfig } from "./utils/web3-onboard/wagmi";
 import walletConnect from "@web3-onboard/walletconnect";
 import { WagmiProvider } from "wagmi";
 import { defineChain, http } from "viem";
@@ -64,7 +64,7 @@ import { defineChain, http } from "viem";
 const queryClient = new QueryClient({
 	defaultOptions: {
 		queries: {
-			refetchInterval: 2000,
+			staleTime: 15_000,
 			queryKeyHashFn: hashQueryKey,
 		},
 	},
@@ -72,6 +72,13 @@ const queryClient = new QueryClient({
 
 
 const projectId = "5ac86584b1de10a7953c6d7b19b52dad";
+
+const injectedWallet = injectedModule();
+
+const walletConnectWallet = walletConnect({
+	projectId,
+	dappUrl: "https://wardenprotocol.org/",
+});
 
 const web3Onboard = init({
 	appMetadata: {
@@ -94,14 +101,7 @@ const web3Onboard = init({
 		rpcUrl: env.evmURL,
 	}],
 	connect: { autoConnectAllPreviousWallet: true },
-	wallets: [injectedModule(), walletConnect({
-		handleUri: (uri: any) => {
-			console.log("wcuri", uri);
-		},
-		projectId,
-		dappUrl: "https://wardenprotocol.org/",
-	})
-	],
+	wallets: [injectedWallet, walletConnectWallet],
 	wagmi,
 });
 
@@ -156,7 +156,14 @@ const chain = defineChain({
 
 const defaultConfig = createConfig({
 	chains: [chain],
-	transports: [http()],
+	transports: {
+		[chain.id]: http(undefined, {
+			batch: {
+				batchSize: 1000,
+				wait: 16
+			},
+		})
+	},
 })
 
 function InjectWagmi({ children }: { children: React.ReactNode }) {
@@ -171,41 +178,14 @@ function InjectWagmi({ children }: { children: React.ReactNode }) {
 
 function App() {
 	const signerOptions: SignerOptions = {};
-	const supportedWallets = wallets.for("keplr", "leap", "cosmostation");
 
 	return (
 		<React.StrictMode>
 			<FaroErrorBoundary>
 				<ChainProvider
-					chains={[
-						...chains,
-						wardenprotocollocal,
-						wardenprotocoldevnet,
-						wardenprotocolbuenavista,
-						wardenprotocolchiado,
-					]}
-					assetLists={[
-						...assets,
-						wardenprotocollocalAssets,
-						wardenprotocoldevnetAssets,
-						wardenprotocolbuenavistaAssets,
-						wardenprotocolchiadoAssets,
-					]}
-					wallets={supportedWallets}
-					walletConnectOptions={{
-						signClient: {
-							projectId: "5ac86584b1de10a7953c6d7b19b52dad",
-							relayUrl: "wss://relay.walletconnect.org",
-							metadata: {
-								name: "Warden Protocol Wallets",
-								description: "Warden Protocol WalletConnect",
-								url: "https://wardenprotocol.org/",
-								icons: [
-									"https://avatars.githubusercontent.com/u/158038121",
-								],
-							},
-						},
-					}}
+					chains={chains}
+					assetLists={assets}
+					wallets={[]}
 					signerOptions={signerOptions}
 				>
 					<Web3OnboardProvider web3Onboard={web3Onboard}>
@@ -213,92 +193,90 @@ function App() {
 							<QueryClientProvider client={queryClient}>
 								<MetaMaskProvider>
 									<AddressProvider>
-										<WalletProvider>
-											<DenomProvider>
-												<BrowserRouter>
-													<FaroRoutes>
-														<Route element={<Root />}>
-															<Route
-																path="/"
-																element={
-																	<DashboardPage />
-																}
-															/>
-															<Route
-																path="/rules"
-																element={
-																	<IntentsPage />
-																}
-															/>
-															<Route
-																path="/staking"
-																element={
-																	<StakingPage />
-																}
-															/>
-															<Route
-																path="/governance"
-																element={
-																	<GovernancePage />
-																}
-															/>
-															<Route
-																path="/actions"
-																element={
-																	<ActionsPage />
-																}
-															/>
-															<Route
-																path="/explorer"
-																element={
-																	<ExplorerPage />
-																}
-															/>
-															<Route
-																path="/explorer/block-by-height/:height"
-																element={
-																	<BlockByHeightPage />
-																}
-															/>
-															<Route
-																path="/keys"
-																element={<KeysPage />}
-															/>
-															<Route
-																path="/keychains"
-																element={
-																	<KeychainsPage />
-																}
-															/>
-															<Route
-																path="/assets"
-																element={<AssetsPage />}
-															/>
-															<Route
-																path="/apps"
-																element={<AppsPage />}
-															/>
-															<Route
-																path="/apps/open"
-																element={
-																	<AppsOpenPage />
-																}
-															/>
-															<Route
-																path="/settings"
-																element={
-																	<SettingsPage />
-																}
-															/>
-															<Route
-																path="/owners"
-																element={<OwnersPage />}
-															/>
-														</Route>
-													</FaroRoutes>
-												</BrowserRouter>
-											</DenomProvider>
-										</WalletProvider>
+										<DenomProvider>
+											<BrowserRouter>
+												<FaroRoutes>
+													<Route element={<Root />}>
+														<Route
+															path="/"
+															element={
+																<DashboardPage />
+															}
+														/>
+														<Route
+															path="/rules"
+															element={
+																<IntentsPage />
+															}
+														/>
+														<Route
+															path="/staking"
+															element={
+																<StakingPage />
+															}
+														/>
+														<Route
+															path="/governance"
+															element={
+																<GovernancePage />
+															}
+														/>
+														<Route
+															path="/actions"
+															element={
+																<ActionsPage />
+															}
+														/>
+														<Route
+															path="/explorer"
+															element={
+																<ExplorerPage />
+															}
+														/>
+														<Route
+															path="/explorer/block-by-height/:height"
+															element={
+																<BlockByHeightPage />
+															}
+														/>
+														<Route
+															path="/keys"
+															element={<KeysPage />}
+														/>
+														<Route
+															path="/keychains"
+															element={
+																<KeychainsPage />
+															}
+														/>
+														<Route
+															path="/assets"
+															element={<AssetsPage />}
+														/>
+														<Route
+															path="/apps"
+															element={<AppsPage />}
+														/>
+														<Route
+															path="/apps/open"
+															element={
+																<AppsOpenPage />
+															}
+														/>
+														<Route
+															path="/settings"
+															element={
+																<SettingsPage />
+															}
+														/>
+														<Route
+															path="/owners"
+															element={<OwnersPage />}
+														/>
+													</Route>
+												</FaroRoutes>
+											</BrowserRouter>
+										</DenomProvider>
 									</AddressProvider>
 								</MetaMaskProvider>
 							</QueryClientProvider>