+++ spaceward/src/features/dashboard/Actions.tsx
@@ -57,7 +57,7 @@ export function Actions() {
 	}
 
 	return (
-		<div className="bg-card  py-5 px-6 mt-6 border-[1px] border-border-edge rounded-2xl">
+		<div className="bg-card py-5 px-6 border-[1px] border-border-edge rounded-2xl">
 			<div className="flex justify-between items-center gap-2 mb-3">
 				<div className="font-bold text-2xl flex items-center justify-between">
 					Latest actions