+++ CHANGELOG.md
@@ -43,6 +43,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Features (non-breaking)
 * (precompiles) Add an ability for contracts to approve actions
 
+### Consensus Breaking Changes
+* (precompiles) Add slinky precompiled contract
+* (x/warden) Add broadcastType to SignRequest
+* (precompiles) Add broadcastType to SignRequest
+
+
 ### Bug Fixes
 
 ## [v0.5.4](https://github.com/warden-protocol/wardenprotocol/releases/tag/v0.5.4) - 2024-11-07